move rabbit message queue to new folder
